# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SIH26034 Legal Metrology Compliance API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/scans/{scan_id}/pair-token": {
            "post": {
                "tags": [
                    "pairing"
                ],
                "summary": "Create Pair Token",
                "operationId": "create_pair_token_scans__scan_id__pair_token_post",
                "parameters": [
                    {
                        "name": "scan_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Scan Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PairTokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pair/{token}/submit": {
            "post": {
                "tags": [
                    "pairing"
                ],
                "summary": "Pair Submit",
                "operationId": "pair_submit_pair__token__submit_post",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Token"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_pair_submit_pair__token__submit_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PairSubmitResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Body_pair_submit_pair__token__submit_post": {
                "properties": {
                    "front": {
                        "type": "string",
                        "format": "binary",
                        "title": "Front"
                    },
                    "back": {
                        "type": "string",
                        "format": "binary",
                        "title": "Back"
                    },
                    "calibration": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "binary"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Calibration"
                    }
                },
                "type": "object",
                "required": [
                    "front",
                    "back"
                ],
                "title": "Body_pair_submit_pair__token__submit_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PairSubmitResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "title": "Ok"
                    },
                    "scanId": {
                        "type": "string",
                        "title": "Scanid"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "ok",
                    "scanId",
                    "status"
                ],
                "title": "PairSubmitResponse"
            },
            "PairTokenResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    },
                    "joinUrl": {
                        "type": "string",
                        "title": "Joinurl"
                    },
                    "expiresAt": {
                        "type": "string",
                        "title": "Expiresat"
                    }
                },
                "type": "object",
                "required": [
                    "token",
                    "joinUrl",
                    "expiresAt"
                ],
                "title": "PairTokenResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}